can't use TIM2

+++ Gesla/BSP/inc/uart_nano.h
@@ -7,7 +7,7 @@
 #include "stm32f10x.h"
 
 extern uint8_t Uart2_DMA_Buffer[ Uart2_NANO_SIZE ];
-extern float chassis_speed_target[3];
+extern int16_t chassis_speed_target[3]; // x y z
 
 void UART_NANO_Init(uint32_t baud);
 void UART_DMA_Init(void);
